refactor(transcription): make the pool storage-agnostic inside app_transcription
The pool and the switchable source move into the app_transcription
package. The pool no longer knows the database: every lifecycle fact
(in progress, transcript, failure, forget, model-switch wipe) is handed
to a TranscriptionStore delegate, and the app wires DriftTranscriptionStore
in the main shell - the single place where pool output meets the drift
transcriptions table. Failure classification (transient vs terminal,
session-guard handling on 401) moves into that store as an application
concern; the store's answer also tells the pool whether to keep
draining its queue. The datasource factory stays app-side since it
maps the white-label app config.
