--- lib/data/transcription/transcription.dart
@@ -1,5 +1,0 @@
-export 'package:app_transcription/app_transcription.dart';
-
-export 'switchable_transcription_source.dart';
-export 'transcription_datasource_factory.dart';
-export 'transcription_service.dart';
