+++ lib/repositories/voicemail/voicemail_repository.dart
@@ -3,11 +3,11 @@ import 'dart:async';
 import 'package:logging/logging.dart';
 
 import 'package:app_database/app_database.dart';
+import 'package:app_transcription/app_transcription.dart';
 
 import 'package:webtrit_api/webtrit_api.dart';
 
 import 'package:webtrit_phone/common/common.dart';
-import 'package:webtrit_phone/data/transcription/transcription.dart';
 import 'package:webtrit_phone/mappers/mappers.dart';
 import 'package:webtrit_phone/models/models.dart';
 import 'package:webtrit_phone/app/session/session.dart';
