+++ packages/data/app_transcription/lib/app_transcription.dart
@@ -2,4 +2,7 @@ export 'package:_http_client/_http_client.dart' show TrustedCertificates;
 
 export 'src/local_whisper_transcription_datasource.dart';
 export 'src/remote_whisper_transcription_datasource.dart';
+export 'src/switchable_transcription_source.dart';
 export 'src/transcription_datasource.dart';
+export 'src/transcription_service.dart';
+export 'src/transcription_store.dart';
